refactor:move styles from screens to components
