--- src/components/Avatar/types.ts
@@ -1,16 +1,0 @@
-export interface AvatarProps {
-  style?: object;
-  id?: string;
-  data:
-    | {
-        url: string;
-      }
-    | null
-    | undefined;
-}
-
-export interface FileEvent {
-  uri: string;
-  name: string;
-  type: string;
-}
