+++ src/components/Elements/SafeArea/types.ts
@@ -1,0 +1,6 @@
+import {ReactNode} from 'react';
+
+export interface SafeAreaProps {
+  children: ReactNode;
+  safeArea?: boolean;
+}
