+++ src/components/Splash/styles.ts
