--- src/components/Wallet/types.ts
@@ -1,3 +1,0 @@
-export interface WalletProps {
-  style?: object;
-}
