--- src/screens/Intro/styles.ts
@@ -1,8 +1,0 @@
-const styles = () => ({
-  screenContainer: {
-    height: '100%',
-    width: '100%',
-  },
-});
-
-export default styles;
